% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrLightGBM.R
\name{LearnerRegrLightGBM}
\alias{LearnerRegrLightGBM}
\alias{mlr_learners_regr.lgbpy}
\title{Regression LightGBM Learner}
\format{[R6::R6Class] inheriting from [mlr3::LearnerRegr].}
\description{
Regression LightGBM Learner

Regression LightGBM Learner
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrLightGBM}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id_col}}{(optional) A character string. The name of the ID column
(default: NULL).}

\item{\code{validation_split}}{A numeric. Ratio to further split the training
data for validation (default: 1). The allowed value range is
0 < validation_split <= 1. This parameter can also be set to
'1', taking the whole training data for validation during the model
training.}

\item{\code{split_seed}}{A integer (default: NULL). Please use this argument in
order to generate reproducible results.}

\item{\code{num_boost_round}}{A integer. The number of boosting iterations
(default: 100).}

\item{\code{early_stopping_rounds}}{A integer. It will stop training if one
metric of one validation data doesn’t improve in last
`early_stopping_round` rounds. '0' means disable (default: 0).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerRegrLightGBM$new()}}
\item \href{#method-train_internal}{\code{LearnerRegrLightGBM$train_internal()}}
\item \href{#method-predict_internal}{\code{LearnerRegrLightGBM$predict_internal()}}
\item \href{#method-importance}{\code{LearnerRegrLightGBM$importance()}}
\item \href{#method-importance2}{\code{LearnerRegrLightGBM$importance2()}}
\item \href{#method-clone}{\code{LearnerRegrLightGBM$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
The initialize function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_internal"></a>}}
\subsection{Method \code{train_internal()}}{
The train_internal function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$train_internal(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_internal"></a>}}
\subsection{Method \code{predict_internal()}}{
The predict_internal function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$predict_internal(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance"></a>}}
\subsection{Method \code{importance()}}{
The importance function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$importance()}\if{html}{\out{</div>}}
}

\subsection{Details}{
A named vector with the learner's variable importances.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance2"></a>}}
\subsection{Method \code{importance2()}}{
The importance2 function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$importance2()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a list with the learner's variable importance values
  and an importance plot.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
