% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.lgbpy-package}
\alias{mlr3learners.lgbpy}
\alias{mlr3learners.lgbpy-package}
\title{mlr3learners.lgbpy: mlr3: LightGBM learner using its python module}
\description{
Brings the LightGBM functionality to the mlr3 framework using the lightgbm python module and reticulate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kapsner/mlr3learners.lgbpy}
  \item Report bugs at \url{https://github.com/kapsner/mlr3learners.lgbpy/issues}
}

}
\author{
\strong{Maintainer}: Lorenz A. Kapsner \email{lorenz.kapsner@gmail.com} (\href{https://orcid.org/0000-0003-1866-860X}{ORCID})

}
